refactor unit convertion function names
