--- src/core/units/fromObject.ts
@@ -1,14 +1,0 @@
-import Unit from './Unit';
-import { Celsius, Fahrenheit } from './Temperature';
-
-export default function unitFromObject(value: number, unit: string): Unit {
-  switch (unit) {
-    // Temperatures
-    case 'celsius':
-      return new Celsius(value);
-    case 'fahrenheit':
-      return new Fahrenheit(value);
-    default:
-      throw new Error(`No suitable unit found for '${unit}'`);
-  }
-}
